This commit replaces CLAP with Pico Args
On my laptop this:
- Reduces binary size by 332 KiB
  - Compiled with RUSTFLAGS='-C codegen-units=1 -C panic=abort -C link-args=-s'
- Reduced time to compile by 3 seconds

I took some creative liberties to add some flair with the help menu and
for the error messages. Command Line Interface should be nearly
identical to the former CLAP version (Error messages are the exception).

All Colors come from ansi_term which was already being pulled in by
CLAP.

Pico was designed to make simple CLI and minimally. It does not
replicate the functionallity of CLAP. If there are changes to the
subcommands, arguements, or flags they will need to be programmed, the
help messaged will need to be updated, and their error message will need
to be made. I would imagine the CLI for pkgar won't change much so I doubt
this will be an issue.