WIP: Refactor pkgar into two crates
This compiles, but the tests fail due to an off by one or similar issue
hiding someplace. Debugging is going to require a significant set of
improvements to the error handling in the crate, the messages are crap
atm.

I also added the thiserror crate to make generating error impls easier,
so it should be fairly straightforward from that side of things to
improve the errors. Note that thiserror isn't no_std right now, so it's
only used in pkgar, pkgar_core implements all the froms without the help
of a macro.