WIP: Refactor extraction into transaction API
The tests fail with this commit, but I want to be able to branch to do
the error refactoring to print out filesystem contents. Might end up
amending this.

Essentially the vec that was internal to the extract function is the
backbone of a new type, Transaction, which provides semi-atomic package
operations against the filesystem. Currently it's basically a refactor
of extract, but it provides a more flexible place to add ugrades and
removes (function headers implemented).

This also entails a large amount of error refactoring to provide helpful
context for various errors. More work to do here.