Implement upgrade; Refactoring; Tests in Rust!
A naive basic upgrade implementation is done now. It iterates a lot and
is currently requesting the entries from the new package once more than
it needs to (Transaction::install is used directly after and calls
read_entries() again). However, it does appear to work, as...

I wrote some tests in rust! The API now has a test case running on it,
so I'll know if I break API or some such. I designed it to be
thread-safe, so parallel testing is possible (unique temp files for each
thread, cleaned up automatically on drop).

This commit also refactors the basic functions in the crate to use
AsRef<Path> instead of &str, and reorganizes the imports in the crate to
make everything look nicer in the rustdoc.