pkgar-keys: Refactor Errors
Fixed two little bugs:
- Private key files now use 600 instead of 644
- Fixed default key file paths

Pulled in the same error deps as the main library,
- UFE for easy pretty-printing of errors
- thiserror for deriving std::error::Error for error types

pkgar-keys now uses two error types, one that includes a file path for
context and an ErrorKind for passing around internally and containing
other error types. I'm still not happy with the sheer volume of
boilerplate needed to create Error, so more work is needed, but error
reporting is much better with these changes, and is probably more
maintainable even given the boilerplate.