Major Error Refactor
This does a similar thing as pkgar_keys, by splitting up the errors into
ErrorKind (which stores the immediate context of the error, for example,
the entry name) and the real Error, which _can_ also have a reason or
filename associated. Error includes a builder pattern for adding context
which can be shorter and is obviously more elegant than what I was doing
before.

Error is also a struct with a private enum member, this is to keep users
from interacting with the error's state, all of which should be accessible
using std::error::Error.