Initial Builder + some tests
pkgar::PackageBuilder provides a reasonably flexible interface for
constructing packages. The API needs a bit of work and there's some
refactoring/other TODOS that need attention, but it seems to work quite
well.

This also includes a bit of hideous test code for format printing a
binary pkgar package for debugging. I'll either delete it or come up
with something much better and include it in a debug module or some
such.