Expand EntryReader; Rearrange imports; Docs
EntryReader now stores a hasher and keeps track of the length of the
data that it has processed, and its state can be checked after it has
been used. This required some modifications to its API in regards to the
trait bounds for most of its impls... This may need further work.

The pkgar crate now re-exports pkgar_keys and pkgar_core so that users
do not need to pull in those libraries indpendently (as their traits are
used extensively in the pkgar API). This may have some impact on
versioning, but I have not done any reading/testing.

There's at least something in terms of documentation now as well.